moved Promise tests to separate file, improved coverage
